src/main/java/com/devops/controller/TodoController.java
package com.devops.controller;

import com.devops.model.Todo;
import com.devops.service.TodoService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;
import org.springframework.web.bind.annotation.RestController;

import javax.validation.Valid;
import java.util.List;
import java.util.UUID;

/**
 * Created by SQLI on 16/11/16.
 */
@RestController
@RequestMapping("/todos")
public class TodoController {

    @Autowired
    private TodoService todoService;

    @RequestMapping(method = RequestMethod.GET, produces = MediaType.APPLICATION_JSON_VALUE)
    public ResponseEntity<List<Todo>> getAll(){
        return ResponseEntity.ok(todoService.getAll());
    }

    @RequestMapping(method = RequestMethod.POST, produces = MediaType.APPLICATION_JSON_VALUE)
    public ResponseEntity createItem(@RequestBody @Valid Todo item){
        UUID id = this.todoService.addItem(item);
        return new ResponseEntity(id, HttpStatus.CREATED);
    }
}
